% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{data}
\name{constants.classic}
\alias{constants.classic}
\title{Newhall classic (monthly inputs) constants}
\format{
A list with 6 elements
\itemize{
\item \code{"depletion.order"} numeric length 64; depletion order
\item \code{"depletion.req"} numeric length 64; depletion requirement
\item \code{"temp.bins"} numeric length 24; temperature bins
\item \code{"pe.bins"} numeric length 24; potential evapotransipration bins
\item \code{"knorth"} matrix 31 x 12
\item \code{"ksouth"} matrix 13 x 12
}
}
\description{
This data set contains constants used for the "classic" (monthly) Newhall model
}
\keyword{datasets}
