% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moistureCondition.R
\name{moistureCondition}
\alias{moistureCondition}
\title{Evaluate soil moisture condition}
\usage{
moistureCondition(soil.profile, n.row = 8, n.col = 8)
}
\arguments{
\item{soil.profile}{numeric vector, typically derived from an 8x8 matrix}

\item{n.row}{integer, number of rows in the soil moisture matrix}

\item{n.col}{integer, number of columns in the soil moisture matrix}
}
\value{
integer representation of soil moisture condition:
\itemize{
\item dry: 1
\item dry/moist: 2
\item moist: 3
}
}
\description{
Evaluate the cells of a soil moisture matrix to determin a soil moisture "condition".
}
