% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newhall.classic.R
\name{Newhall.classic}
\alias{Newhall.classic}
\title{Run Newhall "classic" model using monthly climatic data}
\usage{
Newhall.classic(
  AWC,
  PPT,
  TAVG,
  latitude,
  longitude,
  nsHemisphere,
  ewHemisphere
)
}
\arguments{
\item{AWC}{available water capacity}

\item{PPT}{monthly precipitation (numeric vector length 12)}

\item{TAVG}{monthly temperature (numeric vector length 12)}

\item{latitude}{latitude}

\item{longitude}{longitude}

\item{nsHemisphere}{hemisphere (North/South)}

\item{ewHemisphere}{hemisphere (East/West)}
}
\description{
Run Newhall "classic" model using monthly climatic data
}
